+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.47.1] - 2018-02-27
 ---------------------
 ##### Changed
 - Show separate custom reset messages for drawing, signature and annotation widgets.
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ##### Fixed
 - Date format shown to user in readonly field is different from non-readonly field.
 - Autocomplete widget does not accept options that have multiple subsequent spaces in their label.
+- Draw widget draw color gets reset to the default color after undo-ing strokes.
 
 [4.47.0] - 2018-02-21
 ---------------------